pdf/requirements.txt
PyPDF2>=3.0.0
PyMuPDF>=1.23.0
tqdm>=4.64.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0